@@ -1,116 +1,112 @@
 {
-  "name": "preact-limeapp",
-  "version": "0.1.0-alpha.4",
-  "description": "Limeapp webUI and app",
-  "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --inline --hot --progress",
-    "start": "serve build -s -c 1",
-    "prestart": "npm run build",
-    "build": "cross-env NODE_ENV=production webpack -p --progress --display-optimization-bailout --optimize-minimize && rm ./build/*.map",
-    "prebuild": "mkdirp build && ncp src/assets build/assets",
-    "deploy": "gh-pages -d build",
-    "test": "npm run -s lint && npm run -s test:karma",
-    "test:karma": "karma start test/karma.conf.js --single-run",
-    "lint": "eslint ./",
-    "translations": "node ./node_modules/i18nline/bin/i18nline.js export",
-    "posttranslations": "node ./i18n-diff.js",
-    "cz": "git-cz",
-    "commitmsg": "validate-commit-msg",
-    "release": "standard-version"
-  },
-  "keywords": [
-    "preact",
-    "boilerplate",
-    "webpack"
-  ],
-  "license": "GNU",
-  "author": "Marcos Gutierrez <gmarcos@altermundi.net>",
-  "devDependencies": {
-    "autoprefixer": "^7.1.3",
-    "babel": "^6.23.0",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.0.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-module-resolver": "^2.7.1",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.26.0",
-    "babel-runtime": "^6.26.0",
-    "chai": "^4.1.2",
-    "colors": "^1.1.2",
-    "commitizen": "^2.9.6",
-    "copy-webpack-plugin": "^4.0.1",
-    "core-js": "^2.5.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^0.28.7",
-    "eslint": "^4.5.0",
-    "eslint-config-synacor": "^2.0.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^1.1.5",
-    "gh-pages": "^1.0.0",
-    "html-webpack-plugin": "^2.30.1",
-    "husky": "^0.14.3",
-    "i18nline": "^1.5.0",
-    "isparta-loader": "^2.0.0",
-    "json-loader": "^0.5.7",
-    "karma": "^1.7.1",
-    "karma-chai": "^0.1.0",
-    "karma-chai-sinon": "^0.1.5",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.4",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.4",
-    "less": "^2.7.2",
-    "less-loader": "^4.0.5",
-    "mkdirp": "^0.5.1",
-    "mocha": "^4.0.0",
-    "ncp": "^2.0.0",
-    "offline-plugin": "^4.8.3",
-    "postcss-loader": "^2.0.6",
-    "raw-loader": "^1.0.0-beta.0",
-    "replace-bundle-webpack-plugin": "^1.0.0",
-    "script-ext-html-webpack-plugin": "^1.8.5",
-    "sinon": "^4.0.1",
-    "sinon-chai": "^2.13.0",
-    "source-map-loader": "^0.2.1",
-    "standard-version": "^4.2.0",
-    "style-loader": "^0.19.0",
-    "url-loader": "^0.6.2",
-    "validate-commit-msg": "^2.14.0",
-    "webpack": "^3.5.5",
-    "webpack-dev-server": "^2.7.1"
-  },
-  "dependencies": {
-    "axios": "^0.16.2",
-    "history": "^4.7.2",
-    "jssha": "^2.3.1",
-    "leaflet": "^1.2.0",
-    "leaflet.gridlayer.googlemutant": "^0.6.2",
-    "preact": "^8.2.5",
-    "preact-compat": "^3.17.0",
-    "preact-i18nline": "^1.5.0",
-    "preact-redux": "^2.0.2",
-    "preact-router": "^2.5.7",
-    "preact-router-redux": "^4.1.0",
-    "promise-polyfill": "^6.0.2",
-    "proptypes": "^1.1.0",
-    "purecss": "^1.0.0",
-    "react-load-script": "0.0.6",
-    "react-pure": "^0.3.1",
-    "redux": "^3.7.2",
-    "redux-observable": "^0.17.0",
-    "redux-thunk": "^2.2.0",
-    "rxjs": "^5.4.3",
-    "serve": "^6.0.6",
-    "simple-color-scale": "^1.0.1",
-    "simplegrid": "^1.0.2",
-    "skeleton-less": "^2.0.1",
-    "ulog": "^1.0.3"
-  }
+    "private": true,
+    "name": "preact-limeapp",
+    "version": "0.2.26",
+    "description": "Limeapp webUI and app",
+    "license": "GNU",
+    "author": "Marcos Gutierrez <gmarcos@altermundi.net>",
+    "scripts": {
+        "start": "per-env",
+        "start:production": "npm run -s serve",
+        "start:development": "npm run -s dev",
+        "build": "preact build --no-prerender",
+        "prebuild:production": "lingui extract; lingui compile",
+        "build:production": "preact build --no-prerender --no-sw --no-esm && rm build/*.map",
+        "serve": "preact build && serve build --single",
+        "dev": "preact watch",
+        "lint": "tsc && eslint . --ext .js,.jsx,.ts,.tsx && prettier --ignore-path .gitignore --check \"**/*.+(js|jsx|ts|tsx|json|css|sass|less)\"",
+        "lint:fix": "tsc && eslint . --ext .js,.jsx,.ts,.tsx --fix && prettier --ignore-path .gitignore --write \"**/*.+(js|jsx|ts|tsx|json|css|sass|less)\"",
+        "translations:extract": "lingui extract --clean",
+        "translations:compile": "lingui compile",
+        "cz": "git-cz",
+        "release": "standard-version",
+        "storybook": "start-storybook -h 0.0.0.0 -p 8081",
+        "storybook:build": "build-storybook -c .storybook",
+        "storybook:deploy": "gh-pages -d storybook-static",
+        "test": "jest",
+        "clear-jest": "jest --clearCache",
+        "create-plugin": "node ./devTools/plugins create",
+        "prepare": "husky install"
+    },
+    "devDependencies": {
+        "@babel/preset-react": "^7.18.6",
+        "@lingui/babel-preset-react": "^2.9.2",
+        "@lingui/cli": "^3.14.0",
+        "@lingui/detect-locale": "^3.14.0",
+        "@lingui/loader": "^3.14.0",
+        "@lingui/macro": "^3.14.0",
+        "@storybook/addon-actions": "^6.5.12",
+        "@storybook/addon-controls": "^6.5.12",
+        "@storybook/addon-essentials": "^6.5.12",
+        "@storybook/preact": "^6.5.12",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/preact": "^3.2.2",
+        "@testing-library/user-event": "^14.4.3",
+        "@trivago/prettier-plugin-sort-imports": "^3.3.0",
+        "@types/jest": "^29.2.4",
+        "@types/leaflet": "^1.9.0",
+        "@typescript-eslint/eslint-plugin": "^5.46.0",
+        "@typescript-eslint/parser": "^5.46.0",
+        "babel-jest": "^29.5.0",
+        "babel-plugin-jsx-pragmatic": "^1.0.2",
+        "babel-preset-preact": "^2.0.0",
+        "dotenv": "^16.0.2",
+        "eslint": "^8.24.0",
+        "eslint-config-preact": "^1.3.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-jest": "^27.0.4",
+        "eslint-plugin-prettier": "^4.2.1",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "eslint-plugin-testing-library": "^5.7.0",
+        "fetch-mock": "^9.11.0",
+        "gh-pages": "^4.0.0",
+        "husky": "^8.0.1",
+        "jest": "^27.5.1",
+        "jest-environment-jsdom": "^27.5.1",
+        "jest-extended": "^3.1.0",
+        "jest-fetch-mock": "^3.0.3",
+        "jest-localstorage-mock": "^2.4.22",
+        "jest-preset-preact": "^4.0.5",
+        "less": "^4.1.3",
+        "less-loader": "^7.3.0",
+        "lint-staged": "^13.0.3",
+        "postcss": "^8.4.21",
+        "postcss-less": "^6.0.0",
+        "preact-cli": "^3.4.1",
+        "prettier": "^2.7.1",
+        "standard-version": "^9.5.0",
+        "storybook": "^6.5.12",
+        "tailwindcss": "^3.2.6",
+        "typescript": "^4.9.4",
+        "wait-for-expect": "^3.0.2",
+        "xhr-mock": "^2.5.1"
+    },
+    "dependencies": {
+        "@tanstack/react-query": "^4.6.0",
+        "@tanstack/react-query-devtools": "^4.6.0",
+        "compressorjs": "^1.1.1",
+        "history": "^5.3.0",
+        "leaflet": "^1.9.3",
+        "preact": "^10.11.0",
+        "preact-i18nline": "^2.0.0",
+        "preact-router": "^4.1.0",
+        "react-hook-form": "^7.36.1",
+        "react-leaflet": "^4.2.1",
+        "react-redux": "^8.0.4",
+        "react-router-redux": "^4.0.8",
+        "react-use": "^17.4.0",
+        "redux": "^4.2.0",
+        "redux-observable": "^2.0.0",
+        "simple-color-scale": "^1.0.1",
+        "timeago.js": "^4.0.2"
+    },
+    "husky": {
+        "hooks": {
+            "commit-msg": "commitlint -e $HUSKY_GIT_PARAMS"
+        }
+    },
+    "lint-staged": {
+        "*.{js,ts}": "eslint . --ext .js,.ts --fix",
+        "*.{js,ts,css,less}": "prettier --write"
+    }
 }
