@@ -1,15 +1,12 @@
-export const getMetrics = (api, sid,params) => api.call(sid, 'lime-metrics', 'get_metrics', params)
-	.map(result => Object.assign({}, result, { hostname: params.target })
-	);
+import api from "utils/uhttpd.service";
 
-export const getAllMetrics = (api, sid, params) => params.targets.map(x => getMetrics(api,sid, { target: x }));
+export const getMetrics = (ip) => {
+    console.log("getting metrics for ", ip);
+    return api.call("lime-metrics", "get_metrics", { target: ip });
+};
 
-export const getGateway = (api, sid) => api.call(sid, 'lime-metrics', 'get_gateway', {})
-	.map(result => result);
+export const getGateway = () =>
+    api.call("lime-metrics", "get_gateway", {}).then((res) => res.gateway);
 
-export const getPath = (api, sid, params) => api.call(sid, 'lime-metrics', 'get_path', params)
-	.map(data => Object.keys(data.path).map((key, index) => data.path[key]).reduce((x,y) => x.concat(y), []));
-
-
-export const getLastKnownPath = (api, sid, params) => api.call(sid, 'lime-metrics', 'get_last_internet_path', params)
-	.map(data => Object.keys(data.path).map((key, index) => data.path[key]).reduce((x,y) => x.concat(y), []));
+export const getPath = () =>
+    api.call("lime-metrics", "get_path", {}).then((res) => res.path);
