@@ -1,9 +1,14 @@
-export const getNotes = (api, sid) => api.call(sid, 'lime-utils', 'get_notes', { })
-	.map(x => {
-		if (typeof x.notes === 'undefined') {
-			throw { error: true };
-		}
-		return x;
-	});
+import { map } from "rxjs/operators";
 
-export const setNotes = (api, sid, notes) => api.call(sid, 'lime-utils', 'set_notes', notes);
+export const getNotes = (api) =>
+    api.call("lime-utils", "get_notes", {}).pipe(
+        map((x) => {
+            if (typeof x.notes === "undefined") {
+                throw { error: true };
+            }
+            return x;
+        })
+    );
+
+export const setNotes = (api, notes) =>
+    api.call("lime-utils", "set_notes", notes);
