@@ -1,13 +1,8 @@
-export const getNodeStauts = (api, sid) => api.call(sid, 'lime-utils', 'get_node_status', {});
+import api from "utils/uhttpd.service";
 
-export const getStationTraffic = (api, sid, node) => {
-	if (typeof node === 'undefined') return [];
-	return api.call(sid, 'lime-metrics', 'get_station_traffic', { station_mac: node.station_mac, iface: node.iface })
+export const getNodeStatus = () => {
+    return api.call("lime-utils", "get_node_status", {});
 };
 
-export const getStationSignal = (api, sid, node) => {
-	if (typeof node === 'undefined') return [];
-	return api.call(sid, 'lime-openairview', 'get_station_signal', { station_mac: node.station_mac, iface: node.iface });
-};
-
-export const getInternetStatus = (api, sid) => api.call(sid, 'lime-metrics', 'get_internet_status', {});
+export const getInternetStatus = () =>
+    api.call("lime-metrics", "get_internet_status", {});
