@@ -1,1 +1,7 @@
-module.exports = {};
+module.exports = {
+    plugins: {
+        tailwindcss: {},
+        autoprefixer: {},
+        ...(process.env.NODE_ENV === "production" ? { cssnano: {} } : {}),
+    },
+};
