@@ -1,15 +1,10 @@
-import { combineReducers } from 'redux';
-import { routerReducer } from 'preact-router-redux';
+import { routerReducer } from "react-router-redux";
+import { combineReducers } from "redux";
+import { combineEpics } from "redux-observable";
 
-import { combineEpics } from 'redux-observable';
-
-import createStore from './store/createStore';
-import { loadEpics, loadReducers } from './utils/loader';
-
-import { UhttpdService } from './utils/uhttpd.service';
-
-import { plugins } from './config';
-
+import { plugins } from "./config";
+import createStore from "./store/createStore";
+import { loadEpics, loadReducers } from "./utils/loader";
 
 //GENERATE REDUCERS
 let reducers = loadReducers(plugins);
@@ -17,12 +12,7 @@ reducers.routing = routerReducer;
 const rootReducers = combineReducers(reducers);
 
 //GENERATE EPICS
-const rootEpics =  combineEpics(...loadEpics(plugins));
+const rootEpics = combineEpics(...loadEpics(plugins));
 
 //CREATE STORE
-const store = createStore({},rootEpics,rootReducers, new UhttpdService());
-
-// Init websocket
-store.dispatch({ type: 'meta/CONECTION_START', payload: window.location.href.split('/').slice(0, 3).join('/').concat('/ubus' ) });
-
-export default store;
+export const store = createStore({}, rootEpics, rootReducers);
