@@ -1,82 +1,82 @@
-import { WebSocketService } from '../utils/webSockets.observable';
+import { Observable } from "rxjs";
+import { filter, map } from "rxjs/operators";
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/filter';
+import { WebSocketService } from "../utils/webSockets.observable";
 
 class WebsocketAPI {
-		responses = [];
-		sec = 0;
-		socket;
-		constructor(_wss) {
-			this._wss = _wss;
-			this.jsonrpc = '2.0';
-		}
+    responses = [];
+    sec = 0;
+    socket;
+    constructor(_wss) {
+        this._wss = _wss;
+        this.jsonrpc = "2.0";
+    }
 
-		register(x) {
-    
-			if (x.type === 'message'){
-				let data = JSON.parse(x.data);
-				this.responses[data.id](data);
-				delete this.responses[data.id];
-			}
-			else if (x.type === 'open') {
-				this.responses.conect(true);
-			}
-		}
+    register(x) {
+        if (x.type === "message") {
+            let data = JSON.parse(x.data);
+            this.responses[data.id](data);
+            delete this.responses[data.id];
+        } else if (x.type === "open") {
+            // @ts-ignore
+            this.responses.conect(true);
+        }
+    }
 
-		conect(url) {
-			this.socket = this._wss.connect(url);
-			this.socket.subscribe(this.register.bind(this));
-			let observable = Observable.create((obs) => {
-				this.responses.conect = obs.next.bind(obs);
-			});
-			return observable;
-		}
+    conect(url) {
+        this.socket = this._wss.connect(url);
+        this.socket.subscribe(this.register.bind(this));
+        let observable = new Observable((obs) => {
+            // @ts-ignore
+            this.responses.conect = obs.next.bind(obs);
+        });
+        return observable;
+    }
 
-		addId(){
-			this.sec += 1;
-			return Number(this.sec);
-		}
+    addId() {
+        this.sec += 1;
+        return Number(this.sec);
+    }
 
-		send(mensaje) {
-			const newMensaje = Object.assign({}, mensaje, { id: this.addId(), jsonrpc: this.jsonrpc });
-			this.socket.next(newMensaje);
-			return newMensaje.id;
-		}
+    send(mensaje) {
+        const newMensaje = Object.assign({}, mensaje, {
+            id: this.addId(),
+            jsonrpc: this.jsonrpc,
+        });
+        this.socket.next(newMensaje);
+        return newMensaje.id;
+    }
 
-		call(sid, action, data, method) {
-			let id;
-			let observable = Observable.create((obs) => {
-				if (typeof method === 'undefined') { method = 'call'; }
-				if (method !== 'login')  {
-					id = this.send({ method, params: [
-						sid, '/lime/api', action, data
-					] });
-				}
-				else {
-					id = this.send({ method, params: data });
-				}
+    call(sid, action, data, method) {
+        let id;
+        let observable = Observable.create((obs) => {
+            if (typeof method === "undefined") {
+                method = "call";
+            }
+            if (method !== "login") {
+                id = this.send({
+                    method,
+                    params: [sid, "/lime/api", action, data],
+                });
+            } else {
+                id = this.send({ method, params: data });
+            }
 
-				this.responses[id] = (data) => {
-					if (!data.error) {
-						obs.next(data);
-						obs.complete();
-					}
-					else {
-						obs.error(data.error);
-					}
-
-				};
-    
-			});
-			return observable
-				.filter((x) => x.id === id)
-				.map(x => x.result);
-		}
-
+            this.responses[id] = (data) => {
+                if (!data.error) {
+                    obs.next(data);
+                    obs.complete();
+                } else {
+                    obs.error(data.error);
+                }
+            };
+        });
+        return observable.pipe(
+            filter((x) => x.id === id),
+            map((x) => x.result)
+        );
+    }
 }
-
 
 // Websockets services
 export default new WebsocketAPI(new WebSocketService());
